+++ src/components/form/constants.ts
@@ -1,0 +1,1 @@
+export const SIMPLE_EMAIL_PATTERN = /.+@.+\..+/;
