+++ src/components/form/icons/clear.tsx
