+++ src/page-building/career-details/career-details.stories.mdx
@@ -5,7 +5,7 @@ import {
   Story
 } from '@storybook/addon-docs';
 import { JobDescription } from './job-description';
-import { CareerForm } from '../../components/career-form/career-form';
+import { CareerForm } from './career-form/career-form';
 
 <Meta title="Pages/Career Details" component={{ Image }} />
