+++ src/page-building/career-details/career-details.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Layout } from '../../new-components/layout/layout';
-import { CareerForm } from '../../components/career-form/career-form';
+import { CareerForm } from './career-form/career-form';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
 import { JobDescription } from './job-description';
 import { SyPersonioJob } from '../../types';
