+++ src/page-building/career-details/career-form/career-file-preview.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import React, { ReactNode } from 'react';
-import { Clear } from './icons/clear';
-import { up } from '../../new-components/support/breakpoint';
+import { Clear } from '../../../components/form/icons/clear';
+import { up } from '../../../new-components/support/breakpoint';
 import { rgba } from 'polished';
 
 interface FilePreviewProps {
