+++ src/page-building/career-details/career-form/career-file-upload-type.tsx
@@ -10,8 +10,8 @@ import {
   UseFormWatch,
 } from 'react-hook-form';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
-import { up } from '../../new-components/support/breakpoint';
-import { SelectArrow } from './icons/select-arrow';
+import { up } from '../../../new-components/support/breakpoint';
+import { SelectArrow } from '../../../components/form/icons/select-arrow';
 
 interface CareerFileUploadTypeProps {
   setValue: UseFormSetValue<FieldValues>;
