+++ src/page-building/contact/contact-form.tsx
@@ -1,21 +1,21 @@
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
-import { CheckmarkIcon } from '../career-form/icons/checkmark';
-import { RightArrowIcon } from '../career-form/icons/right-arrow';
+import { CheckmarkIcon } from '../../components/form/icons/checkmark';
+import { RightArrowIcon } from '../../components/form/icons/right-arrow';
 import {
   ButtonText,
   ErrorMessageSend,
+  InputField,
   SendButton,
   SentButton,
   Sup,
-} from './controls';
-import { Grid, GridItem } from '../grid/grid';
-import { InputField } from '../career-form/career-components';
-import { SIMPLE_EMAIL_PATTERN } from '../career-form/career-form';
-import { Link } from '../links/links';
-import { CaptionText } from '../typography/typography';
+} from '../../components/form/controls';
+import { Grid, GridItem } from '../../components/grid/grid';
+import { Link } from '../../components/links/links';
+import { CaptionText } from '../../components/typography/typography';
 import { Trans, useTranslation } from 'gatsby-plugin-react-i18next';
 import styled from 'styled-components';
+import { SIMPLE_EMAIL_PATTERN } from '../../components/form/constants';
 
 type RequestStatus = 'pending' | 'success' | 'error';
